refactored Winlink form support into a separate package 'internal/forms'
